% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_authors.R
\name{abbreviate_authors}
\alias{abbreviate_authors}
\title{Abbreviate Authors}
\usage{
abbreviate_authors(oa_works_df)
}
\arguments{
\item{oa_works_df}{A data frame containing the works. It should have columns 'publication_year' and 'author'.
'author' should be a data frame with a column 'au_display_name'. One example is the data frame from `openalexR`.}
}
\value{
A vector of abbreviated author names and publication years.
}
\description{
This function abbreviates the author names in a given data frame of works.
The output is a list of author names in the format "First Author et al. (Year)" or "Author1 & Author2 (Year)".
}
\examples{
\dontrun{
abbreviate_authors(oa_works_df)
}

}
