% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_read.r
\name{read_corpus}
\alias{read_corpus}
\title{Alias to arrow_read_corpus}
\usage{
read_corpus(path = file.path("data", "corpus"), unify_schemas = FALSE)
}
\description{
This function is an alias to \code{\link{arrow_read_corpus}}.
For more details, see the documentation for \code{\link{arrow_read_corpus}}.
}
