% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R
\name{compact}
\alias{compact}
\title{Compact}
\usage{
compact(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A modified string with newlines, asterisks, extra spaces, and unnecessary spaces around parentheses removed.
}
\description{
This function removes newlines, asterisks, extra spaces, and unnecessary spaces around parentheses from a string.
}
\examples{
x <- "This is a\nstring with *extra* spaces (and unnecessary spaces)."
compact(x)
}
