% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_filter_ids.r
\name{corpus_filter_ids}
\alias{corpus_filter_ids}
\title{Filter corpus by IDs and save as Arrow datasets}
\usage{
corpus_filter_ids(
  arrow_dir = file.path(".", "data", "corpus"),
  arrow_filter_dir = file.path(".", "data", "corpus_filtered"),
  filter_ids = NULL
)
}
\arguments{
\item{arrow_dir}{The directory path where the corpus Arrow datasets are located. default: \verb{file.path(".", "data", "corpus"}}

\item{arrow_filter_dir}{The directory path where the filtered Arrow datasets will be saved. default: \code{file.path(".", "data", "corpus_filtered")}}

\item{filter_ids}{A vector of IDs to filter the corpus by.}
}
\description{
This function filters a corpus of data by a set of IDs and saves the filtered data as Arrow datasets.
}
\details{
The function performs the following steps:
\enumerate{
\item Retrieves the list of year directories within the \code{arrow_dir} path.
\item Extracts the publication years from the directory names.
\item Iterates over each year directory and performs the following operations:
\itemize{
\item Reads the corpus dataset for the current year.
\item Filters the dataset by the provided \code{filter_ids}.
\item If the filtered dataset is not empty, it is mutated to include the publication year and saved as an Arrow dataset in the \code{arrow_filter_dir} path.
}
}
}
