% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_pages_to_arrow.r
\name{corpus_pages_to_arrow}
\alias{corpus_pages_to_arrow}
\title{Convert Downloaded Corpus Data to Arrow Format}
\usage{
corpus_pages_to_arrow(
  pages_dir = file.path(".", "data", "pages"),
  arrow_dir = file.path(".", "data", "corpus"),
  continue = TRUE,
  delete_arrow_dir = FALSE,
  verbose = FALSE,
  dry_run = FALSE,
  mc_cores = 3
)
}
\arguments{
\item{pages_dir}{The directory where the downloaded data is stored.}

\item{arrow_dir}{The directory where the converted Arrow data will be stored.}

\item{continue}{Logical indicating whether to continue from where the conversion was left off.}

\item{delete_arrow_dir}{Logical indicating whether to delete the existing Arrow directory before conversion.}

\item{verbose}{Logical indicating whether to display messages.}

\item{dry_run}{Logical indicating whether to run the function without converting any data. Default is FALSE.}

\item{mc_cores}{The number of cores to be used for parallel processing. Default is 3.
This is limited by memory requirements.}
}
\description{
This function converts downloaded data in RDS format to Arrow format.
}
