% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_read.r
\name{corpus_read}
\alias{corpus_read}
\title{Get Corpus}
\usage{
corpus_read(path = file.path("data", "corpus"), unify_schemas = FALSE)
}
\arguments{
\item{path}{The path to the dataset.}

\item{unify_schemas}{Logical indicating whether to unify the schemas of the datasets.
\code{TRUE} is the safer option, but takes more time to read the first time,Default is FALSE.}
}
\value{
A corpus object.
}
\description{
This function opens a dataset from a specified path and returns a corpus object.
}
\examples{
\dontrun{
corpus_read("data/corpus/")
}

}
