% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_clean.R
\name{doi_clean}
\alias{doi_clean}
\title{Clean DOIs}
\usage{
doi_clean(dois)
}
\arguments{
\item{dois}{Character vector of DOIs to clean.}
}
\value{
Character vector of cleaned DOIs.
}
\description{
Remove resolver URLs from DOIs, keeping the full DOI intact.
\strong{WARNING: This removes everything before the first \code{10.} part of the URL.
This will likely work, onless the URL for the resolver contains \code{10.}
which is rather unlikely}
}
\examples{
dois <- c(
  "https://doi.org/10.1000/xyz123",
  "http://dx.doi.org/10.1016/j.scitotenv.2022.123456",
  "https://handle.net/10.1038/nphys1170",
  "https://resolver.example.com/10.1002/xyz123",
  "https://journals.somepublisher.com/doi/10.1109/5.771073",
  "https://customresolver.university.edu/path/10.1177/0022167819876543",
  "10.1000/xyz123" # Already clean
)
clean_dois(dois)
# [1] "10.1000/xyz123"                "10.1016/j.scitotenv.2022.123456"
# [3] "10.1038/nphys1170"             "10.1002/xyz123"
# [5] "10.1109/5.771073"              "10.1177/0022167819876543"
# [7] "10.1000/xyz123"
}
