% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_retracted.R
\name{doi_not_retracted}
\alias{doi_not_retracted}
\title{Check if DOIs are not retracted}
\usage{
doi_not_retracted(dois, cache_file = NULL, email = NULL)
}
\arguments{
\item{dois}{A vector of DOIs to be validated. Resolver will be removed}

\item{cache_file}{A file name of the cache to be used, i.e. the downloaded retraction data. THe file is an `rds` file as downloaded from the retractionwatch site. If NULL, the data will not be cached.}

\item{email}{An optional email address to be included in the API request [RECOMMENDET!].}
}
\value{
A named logical vector indicating whether each DOI is retracted (`FALSE`) or or not (`TRUE`), names are the dois.
}
\description{
This function checks if a given list of DOIs (Digital Object Identifiers) are retracted.
It uses the Crossref API to query the Retraction Watch database.
}
\details{
The dois are cleaned, i.e. the resolver are removed, before processing.
}
\examples{
# Check if a single DOI is retracted
doi_not_retracted("10.1234/abcd")

# Check if multiple DOIs are retracted
dois <- c("sbcd1234", "10.1234/abcd", "10.1002/jcb.23190", "10.47366/sabia.v5n1a3")
doi_not_retracted(dois)

}
