% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_authors.R
\name{extract_authors}
\alias{extract_authors}
\title{Extract authors' institution country codes}
\usage{
extract_authors(
  data,
  positions = c("first", "middle", "last", NA),
  unserialize = NULL
)
}
\arguments{
\item{data}{A data frame containing authors' information, including the institution country codes.}

\item{positions}{A character vector specifying the positions of the authors to extract.
This has to be one or more of \code{first}, \code{middle}, \code{last} or \code{NA}. If logical \code{TRUE} all authors are extracted.
Default: c("first", "middle", "last", NA). i.e. all authors.}
}
\value{
A \code{tibble} with all columns in the nested \code{data.frame} in \code{data$author[[i]]} plustwo columns:
\itemize{
\item \code{works_id}: the work id
\item \code{publication_year}: the publiction year of the work
}
}
\description{
This function extracts the author data which is nested in the provided \code{data} argument.
}
\details{
The \code{extract_authors} function takes a data frame of with one column named \code{author} containing the auther information
nd returns these as a \code{tibble} with additional info of the OpenAlex \code{id} of the work and the \code{publication_y}.
}
\examples{
authors <- data.frame(
    institution_country_code = c("US", "UK", "DE"),
    stringsAsFactors = FALSE
)
extract_countries(authors)

}
