% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corpus.R
\name{get_corpus}
\alias{get_corpus}
\title{Get Corpus}
\usage{
get_corpus(
  select = c("id", "doi", "authorships", "publication_year", "title",
    "abstract_inverted_index", "topics"),
  verbose = FALSE,
  arrow_directory = "corpus",
  mc_cores = 8,
  ...
)
}
\arguments{
\item{select}{A character vector specifying the columns to select from the data. Default is a set of predefined
columns.}

\item{verbose}{A logical value indicating whether to display verbose output. Default is FALSE.}

\item{arrow_directory}{The path where the parquet files will be saved. Default is "corpus".}

\item{mc_cores}{The number of CPU cores to use for parallel processing. Default is 8.}

\item{...}{Additional filter arguments to be passed to the \code{openalexR::oa_fetch} function.}
}
\value{
path to the arrow directory
}
\description{
This function retrieves data from OpenAlex and saves it to an arrow database using the parquet file format.
The database is partitioned by `publication_year``.
In case of a crash during the download, the function, when ryn again,  will skip the existing files and continue with
the next year.
}
\examples{
\dontrun{
get_corpus(title_and_abstract.search = "biodiversity")
}

}
