% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_count.R
\name{get_count}
\alias{get_count}
\title{FUNCTION_TITLE}
\usage{
get_count(search_term, dois = NULL)
}
\arguments{
\item{search_term}{PARAM_DESCRIPTION}

\item{dois}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openalexR]{oa_query}}, \code{\link[openalexR]{oa_request}}
 \code{\link[dplyr]{select}}
}
