\name{IPBES.R-package}
\alias{IPBES.R-package}
\alias{IPBES.R}
\docType{package}
\title{\packageTitle{IPBES.R}}
\description{\packageDescription{IPBES.R}}
\section{Package Content}{\packageIndices{IPBES.R}}
\author{\packageAuthor{IPBES.R}}
\section{Maintainer}{\packageMaintainer{IPBES.R}}
\keyword{package}
