% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_snapshot_to_arrow.R
\name{oa_snapshot_to_arrow}
\alias{oa_snapshot_to_arrow}
\title{Convert OA snapshot to Arrow format}
\usage{
oa_snapshot_to_arrow(
  snapshot_dir = file.path("", "Volumes", "openalex", "openalex-snapshot"),
  arrow_dir = file.path("", "Volumes", "openalex", "arrow"),
  data_sets = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{snapshot_dir}{The directory path of the OA snapshot data. Default is "Volumes/openalex/openalex-snapshot".}

\item{arrow_dir}{The directory path where the Arrow files will be saved. Default is "Volumes/openalex/arrow".}

\item{data_sets}{A character vector specifying the data sets to process. Default is NULL, which processes all data sets.}
}
\description{
This function converts the OA (openalex) snapshot data to Arrow format.
}
\examples{
\dontrun{
# Convert all data sets in the default snapshot directory
oa_snapshot_to_arrow()

# Convert specific data sets in a custom snapshot directory
oa_snapshot_to_arrow(snapshot_dir = "/path/to/snapshot", data_sets = c("data_set1", "data_set2"))
}

}
