% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snowball.R
\name{plot_snowball}
\alias{plot_snowball}
\title{Plot Snowball}
\usage{
plot_snowball(snowball, name, path = "figures")
}
\arguments{
\item{snowball}{A snowball object containing the data to be plotted.}

\item{name}{The name to be used in the plot titles and file names.}

\item{path}{The path where the plot files will be saved. Default is "figures".}
}
\value{
No return value, called for side effects.
}
\description{
This function takes a snowball object and a name, and creates two plots: one sized by cited_by_count and the other by cited_by_count_by_year.
The plots are saved as both PDF and PNG in the specified path.
}
\examples{
\dontrun{
plot_snowball(snowball, "example")
}
}
