% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snowball_interactive.R
\name{plot_snowball_interactive}
\alias{plot_snowball_interactive}
\title{Plot an Interactive Citation Network from a Snowball Search}
\usage{
plot_snowball_interactive(snowball)
}
\value{
A networkD3 object representing the interactive network plot.
}
\description{
This function creates a interactive snowball seaarch network using the networkD3 package.
}
\examples{
\dontrun{
plot_snowball_interactive(snowball)
}
}
