% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpus.r
\name{read_corpus}
\alias{read_corpus}
\title{Get Corpus}
\usage{
read_corpus(path = file.path("data", "corpus"), unify_schemas = FALSE)
}
\arguments{
\item{path}{The path to the dataset.}
}
\value{
A corpus object.
}
\description{
This function opens a dataset from a specified path and returns a corpus object.
}
\examples{
get_corpus("data/corpus/")
}
