% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_arrow.r
\name{serialize_arrow}
\alias{serialize_arrow}
\title{Serialize Arrow}
\usage{
serialize_arrow(data)
}
\arguments{
\item{data}{A data frame containing 'topics' and 'author' columns.}
}
\value{
A modified data frame with serialized 'topics' and 'author' columns.
}
\description{
This function serializes the 'topics' and 'author' columns of a data frame using base64 encoding.
}
\examples{
data <- data.frame(topics = list("topic1", "topic2"), author = list("author1", "author2"))
serialize_arrow(data)
}
