% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_dt.R
\name{table_dt}
\alias{table_dt}
\title{Create a DataTable with various export options}
\usage{
table_dt(
  x,
  buttons = list(list(extend = "csv", filename = fn), list(extend = "excel", filename =
    fn), list(extend = "pdf", filename = fn, orientation = "landscape", customize =
    DT::JS("function(doc) {", "  doc.defaultStyle.fontSize = 5;", "}")), "print"),
  scroller = TRUE,
  scrollY = JS("window.innerHeight * 0.7 + 'px'"),
  scrollX = TRUE,
  fixedColumns = list(leftColumns = 4),
  escape = FALSE,
  ...
)
}
\arguments{
\item{x}{The data to be displayed in the DataTable}

\item{buttons}{A list of buttons to be displayed in the DataTable}

\item{scroller}{Logical value indicating whether to enable scrolling in the DataTable}

\item{scrollY}{JavaScript code to set the height of the DataTable}

\item{scrollX}{Logical value indicating whether to enable horizontal scrolling in the DataTable}

\item{fixedColumns}{A list specifying the number of fixed columns in the DataTable}

\item{escape}{Logical value indicating whether to escape HTML entities in the DataTable}

\item{...}{Additional options to be passed to the DT::datatable function}
}
\value{
A DataTable object
}
\description{
Create a DataTable with various export options
}
\examples{
table_dt(iris)
}
