% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_corpus.r
\name{test_corpus}
\alias{test_corpus}
\title{Test Corpus}
\usage{
test_corpus(path = file.path("data", "corpus"))
}
\arguments{
\item{path}{The path to the directory containing the parquet files.}
}
\description{
This function tests the parquet files from a specified directory and prints a message for each file.
The function does not return anything, the essential are the disgnostig messages
}
\examples{
test_corpus()

}
\keyword{internal}
