% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_data_frame.R
\name{to_data_frame}
\alias{to_data_frame}
\title{Convert Snowball to result}
\usage{
to_data_frame(snowball)
}
\arguments{
\item{snowball}{A snowball object containing the data to be converted.}
}
\value{
The \code{data.frame} generated
}
\description{
This function takes a snowball object and a filename, and converts the snowball object to an
data.frame containing bibliometric information.
}
\examples{
\dontrun{
to_data_frame(snowball, "example.result")
}
}
