% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unserialize_arrow.r
\name{unserialize_arrow}
\alias{unserialize_arrow}
\title{Unserialize Arrow}
\usage{
unserialize_arrow(x)
}
\arguments{
\item{x}{A character vector containing serialized values, at the moment 'topics' and 'author' columns.}
}
\value{
A vector of lists containing the unseerialu=ized data
}
\description{
This function unserializes the vector `x` and returns a vector containing the extracted values.
}
