% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_download_doi_files.r
\name{zotero_download_doi_files}
\alias{zotero_download_doi_files}
\title{Downloads DOI files from Zotero}
\usage{
zotero_download_doi_files(doi = NULL, dest_dir = NULL)
}
\arguments{
\item{doi}{The DOI (Digital Object Identifier) of the file to be downloaded.}

\item{dest_dir}{The directory where the files will be downloaded.}
}
\description{
This function downloads DOI files from Zotero using the provided DOI.
}
\examples{
\donotrun{
zotero_download_doi_files("10.1234/abcd")
}
}
