% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_get_group.R
\name{zotero_get_group}
\alias{zotero_get_group}
\title{Get Zotero Group Data}
\usage{
zotero_get_group(
  group_id = 2352922,
  file = tempfile(fileext = ".csv"),
  overwrite = FALSE,
  return_data = FALSE,
  api_key = NULL
)
}
\arguments{
\item{group_id}{The ID of the Zotero group.}

\item{file}{The file path to save the retrieved data.}

\item{overwrite}{Logical indicating whether to overwrite an existing file with the same name.}

\item{return_data}{Logical indicating whether to return the data as a data frame.}

\item{api_key}{API key for Zotero. Only needed for private groups.}
}
\value{
If \verb{return_data`` is} TRUE`, the function returns a data frame containing the retrieved data.
Otherwise, it returns the file path where the data is saved.
}
\description{
This function retrieves data from a Zotero group using the Zotero API.
}
\details{
This function uses the Zotero API to retrieve data from a Zotero group. It downloads the data in CSV format
and saves it to a file specified by the \verb{file`` parameter. The function retrieves the data in batches of 100 records at a time and appends them to the file. If }overwrite\verb{ is set to `TRUE}, an existing file with
the same name will be overwritten. The function also provides options to return the data as a data frame and delete
the file after retrieving the data.

For further information on the API see h\href{https://www.zotero.org/support/dev/web_api/v3/start}{ttps://www.zotero.org/support/dev/web_api/v3/start}.
Note: To use this function, you need to have the \pkg{httr2} and \pkg{stringr} packages installed.
}
\examples{
# Retrieve data from Zotero group with ID 2352922 (IPBES IAS Assessment Bibliograhy) and save it
# in the file named "zotero_data.csv"

zotero_get_group(2352922, "zotero_data.csv")

# Retrieve the same data and return it as a as a data frame. The csv file is deleted.

data <- zotero_get_group(2352922, return_data = TRUE, delete_file = TRUE)

}
