% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_get_group_rdf.R
\name{zotero_get_group_rdf}
\alias{zotero_get_group_rdf}
\title{Get Zotero Group Data}
\usage{
zotero_get_group_rdf(
  group_id = 2352922,
  path = tempfile(),
  output_format = "rdf_zotero",
  api_key = NULL
)
}
\arguments{
\item{group_id}{The ID of the Zotero group.}

\item{path}{The path to save the retrieved data.}

\item{output_format}{The format of the rdf file. Supported is \code{rdf_zotero}, \code{rdf_dc} and \code{rdf_bibliontology}
(see \@link[https://www.zotero.org/support/dev/web_api/v3/basics#item_export_formats]{here} for all
supported rdf formats). Default is \code{rdf_zotero}.}

\item{api_key}{API key for Zotero. Only needed for private groups.}
}
\value{
If \code{return_data} The `path`` where the data is saved.
}
\description{
This function retrieves data from a Zotero group using the Zotero API.
}
\details{
This function uses the Zotero API to retrieve data from a Zotero group. It downloads the data in an rdf format
and saves it in batches of 100 records to the folder specified by the \code{path} parameter. The fplder specified will always be deleted before processing.

For further information on the API see h\href{https://www.zotero.org/support/dev/web_api/v3/start}{ttps://www.zotero.org/support/dev/web_api/v3/start}.
Note: To use this function, you need to have the \pkg{httr2} and \pkg{stringr} packages installed.
}
\examples{
# Retrieve data from Zotero group with ID 2352922 (IPBES IAS Assessment Bibliograhy) and save it
# in the file named "zotero_data.csv"

zotero_get_group_rdf(2352922, "zotero_data")

}
