% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_read_corpus.r
\name{arrow_read_corpus}
\alias{arrow_read_corpus}
\alias{read_corpus}
\title{Get Corpus}
\usage{
arrow_read_corpus(path = file.path("data", "corpus"), unify_schemas = FALSE)

read_corpus(path = file.path("data", "corpus"), unify_schemas = FALSE)
}
\arguments{
\item{path}{The path to the dataset.}
}
\value{
A corpus object.
}
\description{
This function opens a dataset from a specified path and returns a corpus object.

This function is an alias to \code{\link{arrow_read_corpus}}.
For more details, see the documentation for \code{\link{arrow_read_corpus}}.
}
\examples{
arrow_get_corpus("data/corpus/")
}
