% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_countries.R
\name{extract_countries}
\alias{extract_countries}
\title{Extract authors' institution country codes}
\usage{
extract_countries(authors, first_only = TRUE, return_all = FALSE)
}
\arguments{
\item{authors}{A data frame containing authors' information, including the institution country codes.}

\item{first_only}{A logical value indicating whether to extract only the first author's institution country code
or for all authors. Default is TRUE.}

\item{return_all}{A logical value indicating whether to return all authors' institution country codes.
The order correponds to the initial vector. Missing countries are \code{NA}. Default is FALSE.}
}
\value{
A `data.frame with two columns:
\itemize{
\item \code{cc}: the country code
\item \code{weight}: the weight of the country code, i.e. \verb{1/number of authors}. 1 if \code{first_author} is `TRUE``.
}
}
\description{
This function extracts the institution country codes of authors from a given data frame.
The \code{extract_authors} function takes a data frame of authors' information and extracts the institution country codes.
By default, it returns a vector containing only the first institution country code for the first author.
If \code{first_only} is set to FALSE, it returns a vector containing the institution country codes for all authors.
If an author has no institution country code or if it is missing, NA is returned.
}
\examples{
\dontrun{
authors <- data.frame(
    institution_country_code = c("US", "UK", "DE"),
    stringsAsFactors = FALSE
)
extract_countries(authors)
}

}
