% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_count.R
\name{get_count}
\alias{get_count}
\title{Get Count}
\usage{
get_count(search_term, dois = NULL, ...)
}
\arguments{
\item{search_term}{The term to search for.}

\item{dois}{A list of DOIs to search within. Default is NULL.}

\item{...}{Additional Additional filter arguments to pass to `oa_query` as}
}
\value{
A named list where the names are the DOIs and the values are the counts of the search term in each DOI.
}
\description{
This function takes a search term and a list of DOIs, and returns the count of the search term in each DOI.
If duplicate DOIs are provided, the function will stop and throw an error.
}
\examples{
\dontrun{
get_count("climate change", c("10.1038/nature12350", "10.1126/science.1259855"))
}
}
