% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_country_codes.R
\name{map_country_codes}
\alias{map_country_codes}
\title{Draw Map Map of Country Codes}
\usage{
map_country_codes(
  data = NULL,
  values = NULL,
  map_type = "countries",
  geodata_path = tempfile()
)
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble} containing at least one column called \code{iso2c} or `iso3c.}

\item{values}{The name of the column to be plotted as the value in the maps. This is also used for the legend title.}

\item{map_type}{A character string specifying the type of map to plot. Must be one of 'countries', 'regions', or 'subregions'. Default is 'countries'.}

\item{geodata_path}{A character string specifying the path to store the geospatial directory to download data to. Default is a temporary file.}
}
\value{
A ggplot object representing the map of the specified map_type.
}
\description{
This function downloads and plots the GADM and IPBES regions according to a list of ISO codes provided.
}
\examples{
\dontrun{
map_country_codes()
map_country_codes(iso3c = c("USA", "CAN"), map_type = "regions")
}

}
