% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snowball.R
\name{plot_snowball}
\alias{plot_snowball}
\title{FUNCTION_TITLE}
\usage{
plot_snowball(snowball, name, path = "figures")
}
\arguments{
\item{snowball}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION, Default: 'figures'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
