% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snowball_interactive.R
\name{plot_snowball_interactive}
\alias{plot_snowball_interactive}
\title{Plot an Interactive Citation Network from a Snowball Search}
\usage{
plot_snowball_interactive(snowball, key_works, file)
}
\arguments{
\item{snowball}{The snowball object containing the network data. The object is returned from the
\link[openalexR]{oa_snowball} function in the `openalexR`` package}

\item{key_works}{A data frame, as returned. e.g. by \verb{oa_fetch(entity = "works", ...},
containing the key-works from the snowball search which will be highlighted in the network.}

\item{file}{The file name to save the network to. TThe directory has tro esxist.
Default: \code{NULL}, i.e. not saved.}
}
\value{
A networkD3 object representing the interactive network plot.
}
\description{
This function creates a interactive snowball seaarch network using the networkD3 package.
}
\examples{
\dontrun{
plot_snowball_interactive(snowball, key_works, file)
}
}
