% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_xlsx.R
\name{to_xlsx}
\alias{to_xlsx}
\title{FUNCTION_TITLE}
\usage{
to_xlsx(snowball, flat_snow, xls_filename)
}
\arguments{
\item{snowball}{PARAM_DESCRIPTION}

\item{flat_snow}{PARAM_DESCRIPTION}

\item{xls_filename}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
}
\seealso{
\code{\link[dplyr]{relocate}}
 \code{\link[writexl]{write_xlsx}}
}
