% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unserialize_arrow.r
\name{unserialize_arrow}
\alias{unserialize_arrow}
\title{Unserialize Arrow}
\usage{
unserialize_arrow(data)
}
\arguments{
\item{data}{A data frame containing serialized 'topics' and 'author' columns.}
}
\value{
A modified data frame with unserialized 'topics' and 'author' columns.
}
\description{
This function unserializes the 'topics' and 'author' columns of a data frame using base64 decoding.
}
\examples{
data <- data.frame(topics = list("dHlwZTE=", "dHlwZTI="), author = list("YXV0aG9yMQ==", "YXV0aG9yMg=="))
unserialize_arrow(data)
}
